Build the standard functions as libraries
Added the possiblity to build the standard functions using a script,
the functions will be built for targets specified in the --target
variable.
The functions will be compiled into a static and dynamic lib containing
both the ST and Rust defined functions.

Added a Dockerfile to build the scripts on debian, still not active